support grouping in reporting
Add support for grouping resulted energy and emission reporting.
By defining a "group" element in the emission lines, the output
includes energy consumption and emission estimate for each group
in addition to the total reported values.

This can be useful in assessing emissions for different sets of
projects, regions or different resource types.